Provide CAMDoomSound tool
It opens doomsound.library and patches its music playback routines
to do true MIDI playback via CAMD. Upon exit it restores the original
library vectors.

For that to work, separate the MUS playback routines out of
doomsound_midi.library into their own musplay.c file
